add a Thrift type for a Blake2 hash
Summary:
Also add conversions to and from the Thrift type to the Rust type.

There are tradeoffs around having parallel sets of data structures vs using
Thrift as the only data structure, but I believe that for these hashes (and
transitively, anything that includes them), having a separate type is best.
That's because:
* these hashes are going to be used all over the place, and using a
heap-allocated vector is just a waste. Even a `SmallVec` that's inline will
consume some extra memory for storing the fact that it's inline.
* something has to verify that the vector is valid, and it's nice for that
to be encoded in the type system.

Reviewed By: farnz

Differential Revision: D7285116

fbshipit-source-id: e887414862821076ef0b11e2d0cb0783be897d4a