Gather a full set of changeset creation timings
Summary:
To diagnose slow changeset creation, we need to know which chunk was
slow. As with upload_blob and process_entries in past diffs, measure key
phases.

The idea is that we can combine all these times (5 per changeset, 2 per blob)
to work out what needs tracing when we're looking for slowness

Reviewed By: StanislavGlebik

Differential Revision: D7195013

fbshipit-source-id: 9769ead981f02d6ecf0258190763e90762936387