ensure that we send takeover data from the correct thread
Summary:
The TakeoverServer::ConnHandler code uses a folly::EventBase to drive the
socket I/O, and therefore must perform all I/O from EventBase thread.
This changes ensures that we always call `sendTakeoverData()` from the correct
thread.

Up until now this has happened to work since the future returned by
`EdenServer::startTakeoverShutdown()` happened to always be completed in the
EdenServer main EventBase thread, and this is the same thread used for the
TakeoverServer socket.

This change makes it explicit that this particular code must be run from the
correct EventBase thread.

Reviewed By: chadaustin

Differential Revision: D7297834

fbshipit-source-id: cbace5f63ea13aa62dd53be0b89d59bea2779cf4