fix detection of unmount in FuseChannel unit tests
Summary:
FuseChannel::processSession() was missing a return statement after we received
EOF from the FUSE device.  This would cause us to fall through and try
processing the empty buffer as a request.  This would fail, and we would try to
send a reply, which would in turn fail since the FUSE device was closed.

This was only an issue with the unit tests: on real FUSE devices we get an
ENODEV error when reading from the device.  In unit tests we get an EOF
instead.

Reviewed By: wez

Differential Revision: D7299336

fbshipit-source-id: c275e03edbf594c0d120257b1cfc89d7c75f60f6