hg: pass the bundle2 caps down to the changegroup packer
Summary:
Previously, during a push the changegroup packer had no information about the
destination server, aside from what changegroup version it supported. In a
future diff we'll want to package certain files and trees based on if the server
is a main server or not, so let's pass thread the bundle2caps down to the
changegroup packer.

The changegroup already had the old style bundlecaps on it, but those
capabilities don't actually exist on the client side during a push, so let's
just use the bundle2 ones, which are more modern anyway.

Differential Revision: D7297899

fbshipit-source-id: 74ab529baccbdc999267555779a10c0d27f8bd47