record the FUSE version reported by the kernel
Summary:
Previously the FuseChannel code unconditionally set `connInfo_.major` and
`connInfo_.minor` to the FUSE version that it supports, rather than the FUSE
version supported by the kernel it is talking to.

This does not appear to be the intended behavior, given that in several other
location it checks the value of `connInfo_->minor` to see if particular
features are supported.

Reviewed By: wez

Differential Revision: D7300845

fbshipit-source-id: 1f5946134037b2cbe03a8b86241acaa401c21ffa