update EdenServer to handle unmounted mounts during takeover
Summary:
As of D7300847 the FuseChannel code will now return the FUSE FD as a closed
File object if the mount was unmounted, even if a takeover stop was requested.
This updates the EdenServer code to deal with the fact that it might receive a
closed FD even during a takeover.  The correct behavior is to simply skip this
mount when building the takeover state to pass to the new process.

This is a somewhat minimal change to fix up this bit of the code.  It would
probably be nice to continue simplifying this logic in the future.

Reviewed By: wez

Differential Revision: D7300877

fbshipit-source-id: c2f617449e155e4b42784ac42ec3aaa378e3288b