Introduce a delaying blob store, for testing
Summary:
We don't want to depend on Manifold being consistent and reliable when
we're trying to fix our performance issues. Introduce a DelayBlob store that
wraps another blobstore and allows you to pass it a function that generates
delay times (e.g. a random number generator with a log-normal distribution).

The idea is that we can use DelayBlob wrapped around RocksBlob to approximate
expected Manifold behaviour, and can improve if needed.

Reviewed By: StanislavGlebik

Differential Revision: D7353228

fbshipit-source-id: fee977d2754f3134af752efc2dbfc77fedc7f2bf