simplecache: refactor to allow just getting or setting, split out jsonserializer
Summary:
I need to be able to test for the presence of a cache without the ability to use a function, and I rather not use a sentinel value or have func raise an exception.

Note that the extension already is vulnerable to race conditions (there is plenty of time between the get, the func() call, and the set as it is).

Also, make it trivial to use JSON as a serializer.

Reviewed By: quark-zju

Differential Revision: D7415721

fbshipit-source-id: dc06e8f3efe725858a7960f7acfecc5a60390b85