mononoke: generate filenodes in blobimport
Summary:
We'll need filenodes in blobimport when we'll add filenodes to the BlobRepo.

The implementation is not great - it creates a separate thread for the
filenodes (see "filenodeinserts"). New filenodes are sent via UnboundedSender
from the parsing cpupool

However it doesn't worth the effort to clean up the code that we are
going to deprecate in a couple of weeks.

Reviewed By: farnz

Differential Revision: D7429440

fbshipit-source-id: 4a9220915bd27f5c1c2028ec604afd700bb8a509