rebase: don't read the collapsemsg if collapse mode is off
Summary:
This one actually broke IMM auto-restarting.

For reasons unknown to humanity, the rebase collapse message is saved (and read) to the same place (.hg/last-message.txt) that the commit message is often written to (before each commit in repo.commitctx(), for example). Thus, `rbsrt.collapsemsg` can sometimes be read from this file incorrectly.

If a rebase fails to commit and you abort, and retry, the abort will read `.hg/last-message.txt` and set the `rebasestate`'s collapsemsg to it. Then, the retry rebase will fail because it has a collapsemsg on its state, but no --collapse.

Anyway, two obvious fixes: let's just only read this message if collapse mode has been set, and let's use a fresh runtime for the non-IMM rebase. I don't think it's safe to use a statefile after abort() has been called on it.

Reviewed By: singhsrb

Differential Revision: D7441659

fbshipit-source-id: f5f8920c3f1d26c8ced083b1c0e375e09fe3248b