remotefilelog: use correct parameter name in peer override
Summary:
`remotefilelog.fileserverclient.peersetup.remotefilepeer` overrides the
`_callstream` method, however it uses `command` rather than `cmd` for the first
parameter name.  This doesn't match the method it's overriding, and clashes
with clienttelemetry's use of this parameter for the original command that the
user ran.

Make this method match all the others.

Reviewed By: quark-zju

Differential Revision: D7443726

fbshipit-source-id: 1170feb21056c3e044bffaf55d95f7c48ff972fb