debugcommands: show fallback colors in debugcolor --style
Summary:
`hg debugcolor --style` shows the component parts of each style individually,
however this doesn't work if the styles are defined as the new fallback styles
(separated by colons).  This is because the fallback is only implemented for
actual style names - it doesn't work for `ui.label('brightblue:blue', 'text')`.

It's usefule to see what the fallbacks are (even if they're not necessary on
your own system), so change debugcolor to split the elements of the fallback
style and show them separately.

Reviewed By: quark-zju

Differential Revision: D7485545

fbshipit-source-id: dce7204c9f0a98bb730b3ba864db28a9ec52a339