datapack: add yieldall (default False) to iterentries
Summary:
A subsequent diff will need access to the node's diff and meta during iteration time. It seems like a
natural part of the API so let's add it.

Note: It's possible to call `.getdelta(name, node)` to get this data if we don't read it here.
But I ran into some weird occassional OSErrors from the mmap API when I did that. So let's just
do this.

Reviewed By: DurhamG

Differential Revision: D7369225

fbshipit-source-id: 252839a549242909153c74287db8f36d6c63bd9c