integration tests: drop run_as_bundle mode
Summary:
The current "run_as_bundle" mode is suboptimal because it causes all
the tests to be run within the same integration runner run. This means that
testpilot can't do things like limit the number of tests run in parallel.

With this change, each test will be run as a separate `integration_runner`
instance.

Thanks to DragonMinded for help figuring this out.

Reviewed By: DragonMinded

Differential Revision: D7493887

fbshipit-source-id: 8d55c2a64573b6bc213dac3daf0d78aa3432d430