update the integration tests to use self.addCleanup()
Summary:
Implement the tear-down actions using addCleanup() rather than through a
tearDown() function.  This simplifies the logic, and also makes it easier to
write the code that can be type-checked more easily.  (We no longer need to
initialize some variables to `None` just to help `tearDown()` identify if they
have been set up yet or not.)

Reviewed By: wez

Differential Revision: D7459278

fbshipit-source-id: de32ebfcb5474a9583f2a51cc9277102189cb77e