fix phabstatus for commit cloud
Summary:
fix `hg ssl` for Commit Cloud

In Commit Cloud world, obstore can sometimes have more knowledge than the unfiltered repo.
For example, in case of many amends on one client, some stuff can not be backed up
and not fetched to another client.
But we don't need to have all obsoleted versions in both repos,
obstore helps us to provide enough information to a user of what has happened
but we need to be carefull with operator []

Reviewed By: DurhamG

Differential Revision: D7516465

fbshipit-source-id: 7e109741829b26c30f20a7528b21b5ebc9b31dde