mercurial-types-mocks: redo the way mock manifests are defined
Summary:
The current structure makes defining new manifests really verbose and
doesn't easily support nested manifests. Make it really simple: either accept a
path map or a description (a csv).

Also switch to a `BTreeMap` to allow lookups in the next diff.

This is unfortunately hard to separate out because all the bits are
interconnected.

Reviewed By: StanislavGlebik

Differential Revision: D7531115

fbshipit-source-id: ca0fa35d4e0ac6a4e23adb2a7d2b5679ce52b643