mononoke: copy-paste select_all method from futures-0.2.0
Summary:
This is a simple but useful combinator that queries lots of streams in parallel
and merges the results.
This is code is taken from futures-0.2.0, but it was modified so that it works
with futures-0.1.*. It was re-formatted with our linter.

The code had a bug that will be fixed in the next diffs

Reviewed By: farnz

Differential Revision: D7550919

fbshipit-source-id: c5b394065c0184a89dfab6a9de699725bc2bd6c2