mononoke: add more tracing
Summary:
Trace parent fetching and linknode fetching for gettreepack

And trace blobstore calls. Note that we are interested in tracing both single
blobstore call and in the retrying manifold calls. Because there can be short
single blobstore calls that fail, but overall retrying manifold call can take a few
seconds.

Reviewed By: farnz

Differential Revision: D7550958

fbshipit-source-id: d607812f558900ee55ddd61b529b7b43e9827f49